Clean up warnings in appwindow. Clean up of various warnings in data provider. Scope property related enums as uint32_t so they can be used in controller interface functions without warnings or having to cast.

+++ src/view/src/rocprofvis_appwindow.h
@@ -90,7 +90,7 @@ private:
     std::unique_ptr<MessageDialog>      m_message_dialog;
     std::unique_ptr<SettingsPanel>      m_settings_panel;
 
-    size_t m_tool_bar_index;
+    int m_tool_bar_index;
     std::function<void(int)> m_notification_callback;
 };
